remove obsolete filetype code
my original theory was that updates are laggy because of syntax
highlighting, which is fixed by blanking the filetype. however, the real
issue turns out to be too frequent buffer updates, which runs in the
main thread, and can cause the entire UI to totally pause while one
update after another are processing. the solution is to not pass so many
updates to the main thread. instead, only update the bufer every ~ 500ms
or so. that way, we can actually keep syntax highlighting on while
streaming.